requirements.txt
# requirements

pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
joblib>=1.1.0
PyYAML>=5.4.1
python-binance>=1.0.16
binance-futures-connector>=1.0.0
